server/requirements.txt
python-docx>=1.1.2
docxtpl>=0.20.0
openpyxl>=3.1.5
pandas>=2.2.3
requests>=2.32.3
fastapi>=0.103.0
uvicorn>=0.29.0
python-multipart>=0.0.20
loguru>=0.7.3
pyyaml>=6.0.1
aiofiles>=24.1.0